pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dqn-rl"
version = "0.1.0"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
    "gymnasium>=0.29.1"
]
requires-python = ">=3.8"
authors = [
    {name = "Mattia Beltrami", email = "your.email@unibo.it"}
]
description = "A modular Deep Q-Network (DQN) package for reinforcement learning"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]